backend/requirements.txt
fastapi
uvicorn
jira
openai
python-dotenv
pydantic
requests
icalendar
python-dateutil
